+++ malachite-nz/src/bin_util/demo_and_bench/gaussian_integer/arithmetic/mod.rs
@@ -16,6 +16,7 @@ pub(crate) fn register(runner: &mut Runner) {
     mul::register(runner);
     neg::register(runner);
     power_of_2::register(runner);
+    shl::register(runner);
     square::register(runner);
     sub::register(runner);
 }
@@ -27,5 +28,6 @@ mod is_power_of_2;
 mod mul;
 mod neg;
 mod power_of_2;
+mod shl;
 mod square;
 mod sub;