+++ malachite-nz/src/test_util/generators/special_random.rs
@@ -276,6 +276,32 @@ pub fn special_random_gaussian_integer_primitive_float_pair_gen<T: PrimitiveFloa
     ))
 }
 
+// -- (GaussianInteger, PrimitiveUnsigned) (small shift amounts) --
+
+pub fn special_random_gaussian_integer_unsigned_pair_gen_var_1<T: PrimitiveUnsigned>(
+    config: &GenConfig,
+) -> It<(GaussianInteger, T)> {
+    Box::new(random_pairs(
+        EXAMPLE_SEED,
+        &|seed| {
+            striped_random_gaussian_integers(
+                seed,
+                config.get_or("mean_stripe_n", 32),
+                config.get_or("mean_stripe_d", 1),
+                config.get_or("mean_bits_n", 64),
+                config.get_or("mean_bits_d", 1),
+            )
+        },
+        &|seed| {
+            geometric_random_unsigneds(
+                seed,
+                config.get_or("mean_small_n", 64),
+                config.get_or("mean_small_d", 1),
+            )
+        },
+    ))
+}
+
 // -- (GaussianInteger, PrimitiveSigned) --
 
 pub fn special_random_gaussian_integer_signed_pair_gen<T: PrimitiveSigned>(